# backend/pkg/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        example: Internal
        type: string
      error:
        example: original server error message
        type: string
      msg:
        example: internal server error
        type: string
      status:
        example: error
        type: string
    type: object
  SuccessResponse:
    properties:
      data:
        type: object
      status:
        example: success
        type: string
    type: object
...
  services.decideToolApprovalRequest:
    properties:
      decision:
        type: string
      edited_args:
        items:
          type: integer
        type: array
      reason:
        type: string
    required:
    - decision
    type: object
...
info:
  contact:
    email: team@pentagi.com
    name: PentAGI Development Team
    url: https://pentagi.com
  description: Swagger API for Penetration Testing Advanced General Intelligence PentAGI.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: PentAGI Swagger API
  version: "1.0"
paths:
...
  /flows/{flowID}/approvals:
    get:
      parameters:
      - description: flow id
        in: path
        minimum: 0
        name: flowID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List a flow's tool approvals
      tags:
      - HITL
  /flows/{flowID}/approvals/{approvalID}/decide:
    post:
      consumes:
      - application/json
      parameters:
      - description: flow id
        in: path
        minimum: 0
        name: flowID
        required: true
        type: integer
      - description: approval id
        in: path
        minimum: 0
        name: approvalID
        required: true
        type: integer
      - description: decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.decideToolApprovalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Decide a pending tool approval
      tags:
      - HITL
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"